requirements.txt
langchain
langchain-core
langchain-tavily
langchain-gigachat
langgraph
pydantic
python-dotenv
langgraph-cli
langgraph-cli[inmem]